fix(콘텐츠 데이터 관리): Content의 watcherCount 타입이 바뀜에 따라 이를 수정합니다.

+++ src/main/java/com/codeit/playlist/domain/content/repository/ContentRepositoryImpl.java
@@ -67,7 +67,7 @@ public class ContentRepositoryImpl implements ContentRepositoryCustom {
                     break;
 
                 case "watcherCount":
-                    Integer cursorWatch = Integer.valueOf(cursor);
+                    long cursorWatch = Integer.valueOf(cursor);
 
                     if (ascending) {
                         builder.and(qContent.watcherCount.gt(cursorWatch)
